moved bootstrapLiquidityPool to graduation.go
