+++ x/iro/keeper/settle.go
@@ -5,11 +5,9 @@ import (
 	"fmt"
 
 	errorsmod "cosmossdk.io/errors"
-	"cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/dymensionxyz/gerr-cosmos/gerrc"
 	"github.com/dymensionxyz/sdk-utils/utils/uevent"
-	"github.com/osmosis-labs/osmosis/v15/x/gamm/pool-models/balancer"
 	gammtypes "github.com/osmosis-labs/osmosis/v15/x/gamm/types"
 
 	"github.com/dymensionxyz/dymension/v3/x/iro/types"
@@ -99,65 +97,6 @@ func (k Keeper) Settle(ctx sdk.Context, rollappId, rollappIBCDenom string) error
 	}
 
 	return nil
-}
-
-// bootstrapLiquidityPool bootstraps the liquidity pool with the raised liquidity and unsold tokens.
-//
-// This function performs the following steps:
-// - Sends the raised liquidity to the IRO module to be used as the pool creator.
-// - Determines the required pool liquidity amounts to fulfill the last price.
-// - Creates a balancer pool with the determined tokens and liquidity.
-// - Uses leftover tokens as incentives to the pool LP token holders.
-func (k Keeper) bootstrapLiquidityPool(ctx sdk.Context, plan types.Plan, poolTokens math.Int) (uint64, sdk.Coins, error) {
-	// claimable amount is kept in the module account and used for user's claims
-	claimableAmt := plan.SoldAmt.Sub(plan.ClaimedAmt)
-
-	// the remaining tokens are used to bootstrap the liquidity pool
-	unallocatedTokens := plan.TotalAllocation.Amount.Sub(claimableAmt)
-
-	// send the raised liquidity token to the iro module as it will be used as the pool creator
-	err := k.BK.SendCoinsFromAccountToModule(ctx, plan.GetAddress(), types.ModuleName, sdk.NewCoins(sdk.NewCoin(plan.LiquidityDenom, poolTokens)))
-	if err != nil {
-		return 0, nil, err
-	}
-
-	denom := plan.SettledDenom
-	if denom == "" {
-		denom = plan.GetIRODenom()
-	}
-
-	// find the raTokens needed to bootstrap the pool, to fulfill last price
-	raTokens, liquidityTokens := types.CalcLiquidityPoolTokens(unallocatedTokens, poolTokens, plan.SpotPrice())
-	rollappLiquidityCoin := sdk.NewCoin(denom, raTokens)
-	baseLiquidityCoin := sdk.NewCoin(plan.LiquidityDenom, liquidityTokens)
-
-	// create pool
-	gammGlobalParams := k.gk.GetParams(ctx).GlobalFees
-	poolParams := balancer.NewPoolParams(gammGlobalParams.SwapFee, gammGlobalParams.ExitFee, nil)
-	balancerPool := balancer.NewMsgCreateBalancerPool(k.AK.GetModuleAddress(types.ModuleName), poolParams, []balancer.PoolAsset{
-		{
-			Token:  baseLiquidityCoin,
-			Weight: math.OneInt(),
-		},
-		{
-			Token:  rollappLiquidityCoin,
-			Weight: math.OneInt(),
-		},
-	}, "")
-
-	// we call the pool manager directly, instead of the gamm keeper, to avoid the pool creation fee
-	poolId, err := k.pm.CreatePool(ctx, balancerPool)
-	if err != nil {
-		return 0, nil, err
-	}
-
-	// calc incentives from leftovers
-	incentives := sdk.NewCoins(
-		sdk.NewCoin(baseLiquidityCoin.Denom, poolTokens.Sub(baseLiquidityCoin.Amount)),
-		sdk.NewCoin(rollappLiquidityCoin.Denom, unallocatedTokens.Sub(rollappLiquidityCoin.Amount)),
-	)
-
-	return poolId, incentives, nil
 }
 
 // addIncentivesToPool adds incentives to the pool
